requirements.txt
Flask
requests
openai
deep-translator
gspread
oauth2client
python-dotenv
